+++ api/send_handler.go
@@ -11,12 +11,14 @@ func sendHandler(c *gin.Context) {
 	connId := c.Query("id")
 	user := c.Query("user")
 	session := c.Query("session")
+	channel := c.Query("channel")
 
-	// Get all worker IDs that the target is connected to
+	// Get all worker IDs that the target(s) is connected to
 	workerIds, apiError := resolveWorkers(common.ResolveOptions{
 		Connection: connId,
 		User:       user,
 		Session:    session,
+		Channel:    channel,
 	})
 	if apiError != nil {
 		apiError.Send(c)
@@ -48,11 +50,14 @@ func sendHandler(c *gin.Context) {
 
 	// Prepare message for worker
 	message := &protos.Message{
-		Body:       body,
-		User:       user,
-		Connection: connId,
-		Session:    c.Query("session"),
-		Type:       parsedMessageType,
+		Type: parsedMessageType,
+		Body: body,
+		Target: &protos.Target{
+			Connection: connId,
+			User:       user,
+			Session:    session,
+			Channel:    channel,
+		},
 	}
 
 	// Send to all workers