+++ common/options.go
@@ -4,7 +4,9 @@ import (
 	"fmt"
 	"github.com/go-redis/redis/v7"
 	"github.com/spf13/viper"
+	"log"
 	"os"
+	"strings"
 )
 
 type JwtOptions struct {
@@ -20,6 +22,8 @@ type DSockOptions struct {
 	Token string
 	/// JWT parsing/verifying options
 	Jwt JwtOptions
+	/// Default channels to subscribe on join
+	DefaultChannels []string
 }
 
 func SetupConfig() {
@@ -35,6 +39,7 @@ func SetupConfig() {
 	viper.SetDefault("redis_password", "")
 	viper.SetDefault("redis_db", 0)
 	viper.SetDefault("address", ":6241")
+	viper.SetDefault("default_channels", "")
 	viper.SetDefault("token", "")
 	viper.SetDefault("jwt_secret", "")
 	viper.SetDefault("debug", false)
@@ -43,7 +48,7 @@ func SetupConfig() {
 
 	if err != nil {
 		if _, ok := err.(viper.ConfigFileNotFoundError); ok {
-			println("No config found, writing new config.")
+			log.Println("No config found, writing new config.")
 
 			err = viper.SafeWriteConfigAs("config.toml")
 
@@ -79,5 +84,8 @@ func GetOptions() DSockOptions {
 		Jwt: JwtOptions{
 			JwtSecret: viper.GetString("jwt_secret"),
 		},
+		DefaultChannels: UniqueString(RemoveEmpty(
+			strings.Split(viper.GetString("default_channels"), ","),
+		)),
 	}
 }