+++ common/resolve_options.go
@@ -4,4 +4,5 @@ type ResolveOptions struct {
 	Connection string
 	User       string
 	Session    string
+	Channel    string
 }