+++ e2e/send_test.go
@@ -30,7 +30,9 @@ func (suite *SendSuite) TestUserSend() {
 	defer conn.Close()
 
 	message, err := sendMessage(sendOptions{
-		User:    "send_user",
+		target: target{
+			User: "send_user",
+		},
 		Type:    "text",
 		Message: []byte("Hello world!"),
 	})
@@ -69,8 +71,10 @@ func (suite *SendSuite) TestUserSessionSend() {
 	defer conn.Close()
 
 	message, err := sendMessage(sendOptions{
-		User:    "send",
-		Session: "session",
+		target: target{
+			User:    "send",
+			Session: "session",
+		},
 		Type:    "text",
 		Message: []byte("Hello world!"),
 	})
@@ -109,8 +113,10 @@ func (suite *SendSuite) TestConnectionSend() {
 	defer conn.Close()
 
 	info, err := getInfo(infoOptions{
-		User:    "send",
-		Session: "connection",
+		target: target{
+			User:    "send",
+			Session: "connection",
+		},
 	})
 	if !checkRequestError(suite.Suite, err, info, "getting info") {
 		return
@@ -124,7 +130,9 @@ func (suite *SendSuite) TestConnectionSend() {
 	id := infoConnections[0].(map[string]interface{})["id"].(string)
 
 	message, err := sendMessage(sendOptions{
-		Id:      id,
+		target: target{
+			Id: id,
+		},
 		Type:    "binary",
 		Message: []byte{1, 2, 3, 4},
 	})
@@ -148,16 +156,60 @@ func (suite *SendSuite) TestConnectionSend() {
 
 func (suite *SendSuite) TestSendNoTarget() {
 	message, err := sendMessage(sendOptions{})
-	if !checkRequestNoError(suite.Suite, err, message, "MISSING_CONNECTION_OR_USER", "sending") {
+	if !checkRequestNoError(suite.Suite, err, message, "MISSING_TARGET", "sending") {
 		return
 	}
 }
 
 func (suite *SendSuite) TestSendNoType() {
 	message, err := sendMessage(sendOptions{
-		Id: "a",
+		target: target{
+			Id: "a",
+		},
 	})
 	if !checkRequestNoError(suite.Suite, err, message, "INVALID_MESSAGE_TYPE", "sending") {
+		return
+	}
+}
+
+func (suite *SendSuite) TestConnectionChannel() {
+	claim, err := createClaim(claimOptions{
+		User:     "send",
+		Session:  "channel",
+		Channels: []string{"send_channel"},
+	})
+	if !checkRequestError(suite.Suite, err, claim, "claim creation") {
+		return
+	}
+
+	conn, resp, err := websocket.DefaultDialer.Dial("ws://worker/connect?claim="+claim["claim"].(map[string]interface{})["id"].(string), nil)
+	if !checkConnectionError(suite.Suite, err, resp) {
+		return
+	}
+
+	defer conn.Close()
+
+	message, err := sendMessage(sendOptions{
+		target: target{
+			Channel: "send_channel",
+		},
+		Type:    "text",
+		Message: []byte("Hello world!"),
+	})
+	if !checkRequestError(suite.Suite, err, message, "sending") {
+		return
+	}
+
+	messageType, data, err := conn.ReadMessage()
+	if !suite.NoError(err, "Error during receiving message") {
+		return
+	}
+
+	if !suite.Equal(websocket.TextMessage, messageType, "Incorrect message type") {
+		return
+	}
+
+	if !suite.Equal([]byte("Hello world!"), data, "Incorrect message data") {
 		return
 	}
 }