+++ worker/send_handler.go
@@ -5,12 +5,13 @@ import (
 	"github.com/Cretezy/dSock/common/protos"
 )
 
-func send(message *protos.Message) {
+func handleSend(message *protos.Message) {
 	// Resolve all local connections for message target
 	connections, ok := resolveConnections(common.ResolveOptions{
-		Connection: message.Connection,
-		User:       message.User,
-		Session:    message.Session,
+		Connection: message.Target.Connection,
+		User:       message.Target.User,
+		Session:    message.Target.Session,
+		Channel:    message.Target.Channel,
 	})
 
 	if !ok {